refactor: create component/form folder
